Improve page load performance across the board.
This PR solves 2 problems —

Problem 1
---------

The `ApplicationController#populate_health_messages` filter was called
on every page load, and was re-computing all health state messages.
This computation had n^2 complexity for every health state message that
had to be rendered, this health state was rendered at the bottom of
every page inside go.

We now don't render the errors on page load, but rely on the ajax poll
to eventually render the errors in a few seconds after the page loads

Problem 2
---------

To further reduce the cost of computation, we now pre-compute the health
state messages, and the pipelines that each server health state message
belongs to, via a background thread.